Refactor networking APIs.
